Remove one unused file by movign method to the other, to be removed before release
